benchmark/requirements.txt
httpx>=0.27.0
pydantic>=2.0.0
pandas>=2.0.0
jinja2>=3.1.0
tiktoken>=0.7.0
python-dotenv>=1.0.0
